🚚 move raw dash types to your own file
